+++ tests/acceptance/resolve-guid-test.ts
@@ -29,7 +29,6 @@ module('Acceptance | resolve-guid', hooks => {
     setupMirage(hooks);
 
     test('User | Index', async assert => {
-        server.create('root', { currentUser: null });
         const user = server.create('user');
 
         await visit(`/${user.id}`);
@@ -39,7 +38,6 @@ module('Acceptance | resolve-guid', hooks => {
     });
 
     test('File | Index', async assert => {
-        server.create('root', { currentUser: null });
         const file = server.create('file', { user: server.create('user') });
 
         await visit(`/${file.id}`);
@@ -55,7 +53,6 @@ module('Acceptance | resolve-guid', hooks => {
         });
 
         test('Index', async assert => {
-            server.create('root', { currentUser: null });
             const node = server.create('node');
 
             await visit(`/${node.id}`);
@@ -65,7 +62,6 @@ module('Acceptance | resolve-guid', hooks => {
         });
 
         test('Forks', async assert => {
-            server.create('root', { currentUser: null });
             const node = server.create('node');
 
             await visit(`/${node.id}/forks`);
@@ -75,7 +71,6 @@ module('Acceptance | resolve-guid', hooks => {
         });
 
         test('Analytics', async assert => {
-            server.create('root', { currentUser: null });
             const node = server.create('node');
 
             await visit(`/${node.id}/analytics`);
@@ -85,7 +80,6 @@ module('Acceptance | resolve-guid', hooks => {
         });
 
         test('Registrations', async assert => {
-            server.create('root', { currentUser: null });
             const node = server.create('node');
 
             await visit(`/${node.id}/registrations`);
@@ -103,7 +97,7 @@ module('Acceptance | resolve-guid', hooks => {
 
         module('No ember_registries_detail_page', __ => {
             test('Index', async assert => {
-                server.create('root', { currentUser: null, activeFlags: [] });
+                server.create('root', 'oldRegistrationDetail');
                 const reg = server.create('registration');
 
                 await visit(`/${reg.id}`);
@@ -113,7 +107,7 @@ module('Acceptance | resolve-guid', hooks => {
             });
 
             test('Forks', async assert => {
-                server.create('root', { currentUser: null, activeFlags: [] });
+                server.create('root', 'oldRegistrationDetail');
                 const reg = server.create('registration');
 
                 await visit(`/${reg.id}/forks`);
@@ -123,7 +117,7 @@ module('Acceptance | resolve-guid', hooks => {
             });
 
             test('Analytics', async assert => {
-                server.create('root', { currentUser: null, activeFlags: [] });
+                server.create('root', 'oldRegistrationDetail');
                 const reg = server.create('registration');
 
                 const url = `/${reg.id}/analytics`;
@@ -137,7 +131,6 @@ module('Acceptance | resolve-guid', hooks => {
 
         module('With ember_registries_detail_page', __ => {
             test('Index', async assert => {
-                server.create('root', { currentUser: null });
                 const reg = server.create('registration');
 
                 await visit(`/${reg.id}`);
@@ -147,7 +140,6 @@ module('Acceptance | resolve-guid', hooks => {
             });
 
             test('Forks', async assert => {
-                server.create('root', { currentUser: null });
                 const reg = server.create('registration');
 
                 await visit(`/${reg.id}/forks`);
@@ -157,7 +149,6 @@ module('Acceptance | resolve-guid', hooks => {
             });
 
             test('Analytics', async assert => {
-                server.create('root', { currentUser: null });
                 const reg = server.create('registration');
 
                 const url = `/${reg.id}/analytics`;
@@ -171,8 +162,6 @@ module('Acceptance | resolve-guid', hooks => {
     });
 
     test('Not found', async assert => {
-        server.create('root', { currentUser: null });
-
         const testCases = [
             { url: '/decaf', test: 'Nonexistant GUID' },
             { url: '/decaf/files', test: 'Nonexistant GUID with existing sub route' },