+++ tests/engines/registries/acceptance/overview/index-test.ts
@@ -29,11 +29,10 @@ module('Registries | Acceptance | overview.index', hooks => {
     setupMirage(hooks);
 
     hooks.beforeEach(function(this: OverviewTestContext) {
-        server.create('root', { currentUser: null });
         server.loadFixtures('registration-schemas');
         this.set('registration', server.create('registration', {
             registrationSchema: server.schema.registrationSchemas.find('prereg_challenge'),
-        }, 'withRegisteredMeta', 'withContributors'));
+        }, 'withContributors'));
     });
 
     test('it renders', async function(this: OverviewTestContext, assert: Assert) {