+++ tests/integration/components/dashboard-item/component-test.ts
@@ -1,19 +1,26 @@
 import { render } from '@ember/test-helpers';
+import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
 import { setupRenderingTest } from 'ember-qunit';
+import { TestContext } from 'ember-test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import { module, test } from 'qunit';
 
-import FakeNode from '../../helpers/fake-node';
-
 module('Integration | Component | dashboard-item', hooks => {
     setupRenderingTest(hooks);
+    setupMirage(hooks);
+
+    hooks.beforeEach(function(this: TestContext) {
+        this.store = this.owner.lookup('service:store');
+    });
 
     test('it renders', async function(assert) {
-        const node = new FakeNode();
-        this.set('node', node);
+        const node = server.create('node', {}, 'withContributors');
+        const project = await this.store.findRecord('node', node.id, { include: 'contributors' });
+
+        this.set('node', project);
         await render(hbs`<DashboardItem @node={{this.node}} />`);
-        assert.dom('div.ember-view > a  > div[class*="_DashboardItem_"]').exists();
-        assert.dom('[data-test-dashboard-item-title]').containsText(node.title, 'Title is set properly');
-        assert.dom(`a[href*="${node.links.html}"]`).exists('Link goes to the node\'s html link');
+        assert.dom('[data-test-dashboard-item-title]').exists();
+        assert.dom('[data-test-dashboard-item-title]').containsText(project.title, 'Title is set properly');
+        assert.dom(`a[href*="${project.links.html}"]`).exists('Link goes to the node\'s html link');
     });
 });