+++ CHANGELOG.md
@@ -28,8 +28,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - Components:
     - `global-link-to`
+    - `<ContributorList>` - modified to take in different parameters
+        - `node` (required)
+        - `shouldTruncate` (default true)
+        - `shouldLinkUsers` (default false)
 
-## [19.1.0]
+## [19.1.0] - 2019-01-23
 ### Added
 - Addons:
     - `loaders-css` - For animated loading indicators
@@ -42,6 +46,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `settings.account`
 - Styles:
     - All styles from `osf-style`
+- Validators:
+    - `httpUrl` - validates that a string looks like an http url
 
 ### Changed
 - Addons:
@@ -51,6 +57,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `ember-cli-sass@8.0.1`
 - Components:
     - `contributor-list` - takes an optional parameter `truncated`
+- Models:
+    - `developer-app` - use custom `httpUrl` validator for urls
 - Templates:
     - `no-implicit-this` template rule activated
 - Types: